WIP see if automatic activation via venv works
At the moment it is not working. The bash seem to only source the .bashrc
file when launched as login shell. Otherwise it is ignored.

Still I think there may be a better way of doing this via entrypoints?
Customizing the runtime environment via variables is known to work in docker.